+++ cypress/integration/all.js
@@ -1,4 +1,24 @@
 describe('All', () => {
+  it('Get random workout', () => {
+    cy.visit('/');
+
+    cy.get('.card-header').find('h3').then(($firstElement) => {
+      let first;
+
+      first = $firstElement.text();
+
+      cy.get('button[name="newWorkout"]').click();
+
+      let second;
+
+      cy.get('.card-header').find('h3').then(($secondElement) => {
+        second = $secondElement.text();
+
+        expect(first).not.to.eq(second);
+      });
+    });
+  });
+
   it('Find random workout and try to start it', () => {
     cy.visit('/');